moves constants
