+++ services/catalog/src/simcore_service_catalog/_constants.py
