+++ services/catalog/src/simcore_service_catalog/api/rest/_services_ports.py
@@ -5,12 +5,12 @@ from fastapi import APIRouter, Depends
 from models_library.api_schemas_catalog.services_ports import ServicePortGet
 from models_library.services_metadata_published import ServiceMetaDataPublished
 
+from ..._constants import RESPONSE_MODEL_POLICY
 from ..dependencies.services import (
     AccessInfo,
     check_service_read_access,
     get_service_from_manifest,
 )
-from ._constants import RESPONSE_MODEL_POLICY
 
 _logger = logging.getLogger(__name__)
