+++ services/catalog/src/simcore_service_catalog/api/rest/_services_specifications.py
@@ -9,12 +9,12 @@ from models_library.api_schemas_catalog.services_specifications import (
 from models_library.services import ServiceKey, ServiceVersion
 from models_library.users import UserID
 
+from ..._constants import RESPONSE_MODEL_POLICY
 from ...db.repositories.groups import GroupsRepository
 from ...db.repositories.services import ServicesRepository
 from ...services.function_services import is_function_service
 from ..dependencies.database import get_repository
 from ..dependencies.services import get_default_service_specifications
-from ._constants import RESPONSE_MODEL_POLICY
 
 router = APIRouter()
 _logger = logging.getLogger(__name__)
